refactor(dto): change salary field type to double in EmployeeDTO"

+++ src/main/java/com/TestingApp/dto/EmployeeDTO.java
@@ -7,6 +7,6 @@ public class EmployeeDTO {
     private  Long id;
     private String email;
     private String name;
-    private  Long salary;
+    private  double salary;
 
 }
